update: make table bar fixed
