+++ console/packages/starwhale-ui/src/base/data-table/headers/header-bar.tsx
@@ -1,0 +1,128 @@
+import * as React from 'react'
+import { StatefulPopover, PLACEMENT, TRIGGER_TYPE } from 'baseui/popover'
+import { StatefulMenu } from 'baseui/menu'
+import IconFont from '../../../IconFont'
+import Button from '../../../Button'
+import { LocaleContext } from 'baseui/locale'
+import { DataTableLocaleT } from '../locale'
+import { IGridState } from '../../../GridTable/types'
+import { useStore } from '../../../GridTable/hooks/useStore'
+import useGrid from '../../../GridTable/hooks/useGrid'
+
+const selector = (s: IGridState) => ({
+    queryinline: s.queryinline,
+    columnleinline: s.columnleinline,
+    onCurrentViewColumnsChange: s.onCurrentViewColumnsChange,
+    wrapperRef: s.wrapperRef,
+    sortable: s.sortable,
+})
+
+function HeaderBar(props: { wrapperWidth: any }) {
+    // @ts-ignore
+    const locale: { datatable: DataTableLocaleT } = React.useContext(LocaleContext)
+    const { wrapperRef, queryinline, columnleinline } = useStore(selector)
+    const { renderConfigQueryInline, renderConfigColumns } = useGrid()
+    const [isShowQuery, setIsShowQuery] = React.useState(false)
+    const [isShowConfigColumns, setIsShowConfigColumns] = React.useState(false)
+
+    const COLUMN_OPTIONS = React.useMemo(
+        () =>
+            [
+                queryinline && {
+                    label: locale.datatable.columnQuery,
+                    type: 'query',
+                },
+                columnleinline && { label: locale.datatable.columnConfig, type: 'column' },
+            ].filter(Boolean),
+        [queryinline, locale]
+    )
+
+    const handleColumnOptionSelect = React.useCallback(
+        (option: any) => {
+            if (option.type === 'query') {
+                setIsShowQuery(true)
+            } else if (option.type === 'column') {
+                setIsShowConfigColumns(true)
+            }
+        },
+        [props]
+    )
+
+    if (!columnleinline && !queryinline) {
+        return <p className='w-30px' />
+    }
+
+    return (
+        <>
+            <div>
+                {renderConfigQueryInline({
+                    width: props.wrapperWidth,
+                    isOpen: isShowQuery,
+                    setIsOpen: setIsShowQuery as any,
+                    mountNode: wrapperRef?.current,
+                })}
+                {renderConfigColumns({
+                    isAction: false,
+                    isOpen: isShowConfigColumns,
+                    setIsOpen: setIsShowConfigColumns as any,
+                    mountNode: wrapperRef?.current,
+                })}
+            </div>
+            <StatefulPopover
+                focusLock
+                triggerType={TRIGGER_TYPE.hover}
+                placement={PLACEMENT.bottom}
+                content={({ close }) => (
+                    <StatefulMenu
+                        items={COLUMN_OPTIONS}
+                        onItemSelect={({ item }) => {
+                            handleColumnOptionSelect(item)
+                            close()
+                        }}
+                        overrides={{
+                            List: { style: { height: '130px', width: '150px' } },
+                            Option: {
+                                props: {
+                                    getItemLabel: (item: { label: string; type: string }) => {
+                                        const icon = {
+                                            query: <IconFont type='filter' />,
+                                            column: <IconFont type='setting' />,
+                                        }
+
+                                        return (
+                                            <div className='flex gap-10px items-center'>
+                                                {icon?.[item.type as keyof typeof icon]}
+                                                {item.label}
+                                            </div>
+                                        )
+                                    },
+                                },
+                            },
+                        }}
+                    />
+                )}
+            >
+                {/* usesd for popover postion ref  */}
+                <Button as='link'>
+                    <div
+                        style={{
+                            alignItems: 'center',
+                            marginLeft: 'auto',
+                            right: 0,
+                            top: -6,
+                            display: 'flex',
+                            width: '30px',
+                            justifyContent: 'center',
+                            marginRight: '8px',
+                        }}
+                    >
+                        <IconFont type='more' />
+                    </div>
+                </Button>
+            </StatefulPopover>
+        </>
+    )
+}
+
+export { HeaderBar }
+export default HeaderBar
