+++ console/packages/starwhale-ui/src/base/data-table/headers/headers.tsx
@@ -9,6 +9,7 @@ import { useStore, useStoreApi } from '@starwhale/ui/GridTable/hooks/useStore'
 import { IGridState } from '@starwhale/ui/GridTable/types'
 import useGridQuery from '@starwhale/ui/GridTable/hooks/useGridQuery'
 import useGrid from '@starwhale/ui/GridTable/hooks/useGrid'
+import HeaderBar from './header-bar'
 
 const sum = (ns: number[]): number => ns.reduce((s, n) => s + n, 0)
 
@@ -175,6 +176,21 @@ export default function Headers({ width }: { width: number }) {
                 width,
             }}
         >
+            <div
+                className='table-headers-bar'
+                style={{
+                    position: 'absolute',
+                    left: '13px',
+                    zIndex: 51,
+                    borderLeftWidth: '0',
+                    overflow: 'visible',
+                    breakInside: 'avoid',
+                    display: 'flex',
+                    height: HEADER_ROW_HEIGHT,
+                }}
+            >
+                {!ctx.isSelectable && <HeaderBar wrapperWidth={width} />}
+            </div>
             <div
                 className='table-headers-pinned'
                 style={{
